[common][feat] APIUtil 삭제
